backend/core/security.py
from datetime import datetime, timedelta, timezone
from jose import JWTError, jwt # pyright: ignore[reportMissingModuleSource]
from typing import Optional
import bcrypt # type: ignore
from .config import settings
from fastapi.security import HTTPAuthorizationCredentials
import models
from sqlalchemy.orm import Session

def verify_password(plain_password: str, hashed_password: str) -> bool:
    return bcrypt.checkpw(plain_password.encode('utf-8')[:72], hashed_password.encode('utf-8'))

def get_password_hash(password: str) -> str:
    return bcrypt.hashpw(password.encode('utf-8')[:72], bcrypt.gensalt()).decode('utf-8')

def create_access_token(data: dict, expires_delta: Optional[timedelta] = None) -> str:
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.now(timezone.utc) + expires_delta
    else:
        expire = datetime.now(timezone.utc) + timedelta(minutes=settings.access_token_expire_minutes)
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, settings.secret_key, algorithm=settings.algorithm)

def decode_token(credentials: HTTPAuthorizationCredentials, db: Session) -> Optional[models.User]:
    try:
        payload = jwt.decode(credentials.credentials, settings.secret_key, algorithms=[settings.algorithm])
        email: Optional[str] = payload.get("sub")
        if email is None:
            return None
    except JWTError:
        return None

    return db.query(models.User).filter(models.User.email == email).first()
