backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
pydantic[email, settings]
python-jose[cryptography]
bcrypt
python-multipart